+++ CHANGELOG.md
@@ -11,6 +11,7 @@ didn't alter the remote pipeline execution, and only escaped the local Python pr
 with the proper remote pipeline execution handling, and possibly per-task timeout enabled by [the new kfp feature](https://github.com/kubeflow/pipelines/pull/10481).
 - Assign pipelines to Vertex AI experiments
 - Migrated `pydantic` library to v2
+- Custom dataset that creates Vertex AI artifact
 
 ## [0.11.1] - 2024-07-01